ed25519

Module pkcs8

PKCS#8 private key support.

Implements Ed25519 PKCS#8 private keys as described in RFC8410 Section 7: https://datatracker.ietf.org/doc/html/rfc8410#section-7

SemVer Notes

The pkcs8 module of this crate is exempted from SemVer as it uses a pre-1.0 dependency (the pkcs8 crate).

However, breaking changes to this module will be accompanied by a minor version bump.

Please lock to a specific minor version of the ed25519 crate to avoid breaking changes when using this module.

Re-exports

pub use pkcs8::spki;

Structs

BitStringRef

ASN.1 BIT STRING type.

Document

ASN.1 DER-encoded document.

KeypairBytes

Ed25519 keypair serialized as bytes.

ObjectIdentifier

Object identifier (OID).

PrivateKeyInfo

PKCS#8 PrivateKeyInfo.

PublicKeyBytes

Ed25519 public key serialized as bytes.

SecretDocument

Secret Document type.

Enums

Error

Error type

Constants

ALGORITHM_ID

Ed25519 Algorithm Identifier.

ALGORITHM_OID

Algorithm ObjectIdentifier for the Ed25519 digital signature algorithm (id-Ed25519).

Traits

DecodePrivateKey

Parse a private key object from a PKCS#8 encoded document.

DecodePublicKey

Parse a public key object from an encoded SPKI document.

EncodePrivateKey

Serialize a private key object to a PKCS#8 encoded document.

EncodePublicKey

Serialize a public key object to a SPKI-encoded document.

Type Aliases

Result

Result type