ed25519

Struct AnyRef

pub struct AnyRef<'a> {
    tag: Tag,
    value: &'a BytesRef,
}

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

This is a zero-copy reference type which borrows from the input data.

Technically ANY hasn’t been a recommended part of ASN.1 since the X.209 revision from 1988. It was deprecated and replaced by Information Object Classes in X.680 in 1994, and X.690 no longer refers to it whatsoever.

Nevertheless, this crate defines an ANY type as it remains a familiar and useful concept which is still extensively used in things like PKI-related RFCs.

Fields
tag: Tagvalue: &'a BytesRef
Implementations

impl<'a> AnyRef<'a>

pub const NULL: AnyRef<'a>

AnyRef representation of the ASN.1 NULL type.

pub const fn new(tag: Tag, bytes: &'a [u8]) -> Result<AnyRef<'a>, Error>

Create a new AnyRef from the provided Tag and DER bytes.

Errors

Returns Error with ErrorKind::Length if bytes is too long.

pub fn value(self) -> &'a [u8]

Get the raw value for this AnyRef type as a byte slice.

pub fn header(&self) -> Header

Returns Tag and Length of self.

pub fn decode_as<T>(self) -> Result<T, <T as DecodeValue<'a>>::Error>
where T: Choice<'a> + DecodeValue<'a>,

Attempt to decode this AnyRef type into the inner value.

Errors

Returns T::Error if a decoding error occurred.

pub fn decode_as_encoding<T>( self, encoding: EncodingRules, ) -> Result<T, <T as DecodeValue<'a>>::Error>
where T: Choice<'a> + DecodeValue<'a>,

Attempt to decode this AnyRef type into the inner value.

Errors

Returns T::Error if a decoding error occurred.

pub fn is_null(self) -> bool

Is this value an ASN.1 NULL value?

pub fn sequence<F, T, E>(self, f: F) -> Result<T, E>
where F: FnOnce(&mut SliceReader<'a>) -> Result<T, E>, E: From<Error>,

Attempt to decode this value an ASN.1 SEQUENCE, creating a new nested reader and calling the provided argument with it.

Errors

Returns E in the event an error is returned from F or if a decoding error occurs.